+++ src/searchers/random_search.cc
@@ -33,7 +33,7 @@ namespace cltune {
 // =================================================================================================
 
 // Randomizes the configurations list
-RandomSearch::RandomSearch(const Configurations &configurations, const float fraction):
+RandomSearch::RandomSearch(const Configurations &configurations, const double fraction):
     Searcher(configurations),
     fraction_(fraction) {
   std::random_shuffle(configurations_.begin(), configurations_.end());
